requirements.txt
torch==2.3.0
transformers==4.40.1
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.8.4
seaborn>=0.13.2
tqdm>=4.66.2
scikit-learn>=1.4.2
python-telegram-bot==21.1.1